refactor: split assumptions.ts (810 lines) into 4 focused modules
Split the monolithic assumptions mutation file into:
- assumptions/types.ts (86 lines) — types + Pothos object refs
- assumptions/temp-zero.ts (280 lines) — launchAssumptionsTempZero
- assumptions/order-invariance.ts (380 lines) — launchOrderInvariance
- assumptions/review.ts (98 lines) — reviewOrderInvariancePair

All 13 throw new Error() calls converted to typed errors
(ValidationError, NotFoundError) from @valuerank/shared.

Added integration tests: assumptions.test.ts (5 tests, all pass).

Adversarial reviews:
- Codex: 5/5 PASS
- Gemini: 5/5 PASS
